Move components to _components
