+++ app/_components/positions/index.ts
@@ -1,0 +1,5 @@
+export * from "./scatteredPositions";
+export * from "./jakePositions";
+export * from "./whaleAndDiversPositions";
+export * from "./projectsPositions";
+export * from "./contactMePositions";
