+++ app/_components/positions/scatteredPositions.ts
