+++ app/page.tsx
@@ -1,4 +1,4 @@
-import { Experience } from "./Experience";
+import { Experience } from "./_components/Experience";
 
 export default function Home() {
   return (
